Add some more quaternion funcs, separate to module
